% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{caldana}
\alias{caldana}
\title{Arabidopsis thaliana growth experiment}
\format{
A data.frame having 140 rows and 3 variables:
\describe{
\item{light}{Light levels}
\item{time}{Time of measurement}
\item{compound}{Metabolic compounds}
}
}
\usage{
data(caldana)
}
\description{
A dataset containing 67 metabolites from plants grown under
different light and temperature conditions. This subset of the data contains
only the light effect and time effect for limited conditions, while the full
data also contains gene expressions.
}
\references{
Caldana C, Degenkolbe T, Cuadros-Inostroza A, Klie S, Sulpice R,
Leisse A, et al. High-density kinetic analysis of the metabolomic and
transcriptomic response of Arabidopsis to eight environmental conditions.
Plant J. 2011;67(5):869-884.
}
