% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdanova.R
\name{fdtest}
\alias{fdtest}
\title{Hypothesis test for densely and relgaruly observed functional data}
\usage{
fdtest(
  X,
  alpha = 0.05,
  tau = NULL,
  B = 1000,
  pairs = NULL,
  transform = T,
  K = 50,
  verbose = F
)
}
\arguments{
\item{X}{a matrix (one sample) or a list of matrices (two samples or more), with observations contained in rows; equal spacing is assumed when \code{transform} is TRUE}

\item{alpha}{significance level}

\item{tau}{the decay parameter, automatically selected if set to \code{NULL}}

\item{B}{the number of bootstrap replicates}

\item{pairs}{a matrix with two columns, used when there are more than two populations, each row specifying a pair of populations whose means are compared, and if set to \code{NULL}, comparisons for all pairs are performed}

\item{transform}{TRUE/FALSE, whether to transform the data into frequency domain via Fourier basis}

\item{K}{a positive integer, the number of bais functions for transforming the data when \code{transform} is TRUE}
}
\value{
a list of objects: \code{reject} is a boolean variable indicating whether the null is rejected or not. If the null if reject, \code{rej.pairs} optionally gives the pairs of samples that lead to rejection. \code{pvalue} is also returned.
}
\description{
Test the mean or differences of means of functional data are zero or not
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) synfd::dense.fd(mu=0.05*g, X=synfd::gaussian.process(), n=30, m=50))

# test for the equality of mean vectors with pairs={(1,3),(2,4)}
res <- fdtest(X,alpha=0.05,pairs=matrix(1:4,2,2),tau=c(0.4,0.5,0.6))

# get p-value
res$pvalue 

# get selected tau
res$tau
}
\references{
\insertRef{Lopes2019+}{hdanova}

\insertRef{Lin2020}{hdanova}
}
