% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdanova.R
\name{fourier.basis}
\alias{fourier.basis}
\title{Create a Fourier basis of K functions in [0, 1]}
\usage{
fourier.basis(K, pts = seq(0, 1, length.out = 50))
}
\arguments{
\item{K}{A positive integer specifying the number of eigenfunctions to generate.}

\item{pts}{A vector specifying the time points to evaluate the basis functions.}
}
\value{
A K by len(pts) matrix, each column containing a basis function.
}
\description{
Create a Fourier basis of K functions in [0, 1]
}
\examples{
basis <- create.basis(3)
head(basis)
}
\keyword{internal}
