% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDANOVA-package.R
\docType{package}
\name{HDANOVA-package}
\alias{HDANOVA-package}
\alias{HDANOVA}
\title{High-Dimensional Analysis of Variance}
\description{
Included methods:
\itemize{
\item ASCA+ - Analysis of Variance Simultaneous Component Analysis
\item APCA+ - ANOVA Principal Component Analysis
\item APLS+ - ANOVA Partial Least Squares
\item LiMM-PCA - Linear Mixed Model PCA
\item MSCA - Multilevel Simultaneous Component Analysis
\item PC-ANOVA - Principal Component Analysis of Variance
\item PRC - Principal Response Curves
\item PERMANOVA - Permutation Based MANOVA
}
}
\seealso{
Main methods: \code{\link{asca}}, \code{\link{apca}}, \code{\link{limmpca}}, \code{\link{msca}}, \code{\link{pcanova}}, \code{\link{prc}} and \code{\link{permanova}}.
Workhorse function underpinning most methods: \code{\link{hdanova}}.
Extraction of results and plotting: \code{\link{asca_results}}, \code{\link{asca_plots}}, \code{\link{pcanova_results}} and \code{\link{pcanova_plots}}
}
\author{
\strong{Maintainer}: Kristian Hovde Liland \email{kristian.liland@nmbu.no} (\href{https://orcid.org/0000-0001-6468-9423}{ORCID})

}
\keyword{internal}
