% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.asca.R
\name{model.frame.asca}
\alias{model.frame.asca}
\alias{model.matrix.asca}
\title{Model Frame and Model Matrix for ASCA-like Models}
\usage{
\method{model.frame}{asca}(formula, ...)
}
\arguments{
\item{formula}{The \code{asca} object.}

\item{...}{Not implemented}

\item{object}{The \code{asca} object.}
}
\description{
Extraction functions to retrieve the \code{model.frame} and \code{model.matrix} of an \code{asca} object.
}
