% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcanova_plots.R
\name{pcanova_plots}
\alias{pcanova_plots}
\alias{scoreplot.pcanova}
\alias{loadingplot.pcanova}
\title{PC-ANOVA Result Methods}
\usage{
\method{scoreplot}{pcanova}(object, factor = 1, comps = 1:2, col = "factor", ...)
}
\arguments{
\item{object}{\code{pcanova} object.}

\item{factor}{\code{integer/character} for selecting a model factor.}

\item{comps}{\code{integer} vector of selected components.}

\item{...}{additional arguments to underlying methods.}
}
\value{
The plotting routines have no return.
}
\description{
Various plotting procedures for \code{\link{pcanova}} objects.
}
\details{
Usage of the functions are shown using generics in the examples in \code{\link{pcanova}}.
Plot routines are available as
\code{scoreplot.pcanova} and \code{loadingplot.pcanova}.
}
\references{
Luciano G, Næs T. Interpreting sensory data by combining principal
component analysis and analysis of variance. Food Qual Prefer. 2009;20(3):167-175.
}
\seealso{
TODO Overviews of available methods, \code{\link{multiblock}}, and methods organised by main structure: \code{\link{basic}}, \code{\link{unsupervised}}, \code{\link{pcanova}}, \code{\link{supervised}} and \code{\link{complex}}.
Common functions for computation and extraction of results are found in \code{\link{pcanova_results}}.
}
