% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{permanova}
\alias{permanova}
\title{Permutation Based MANOVA - PERMANOVA}
\usage{
permanova(formula, data, ...)
}
\arguments{
\item{formula}{Model formula accepting a single response matrix and predictors.
See details in \code{\link[vegan]{adonis2}}.}

\item{data}{The data set to analyse.}

\item{...}{Additional arguments to \code{\link[vegan]{adonis2}}.}
}
\value{
An ANOVA table with permutation-based p-values.
}
\description{
Wrapper for the \code{\link[vegan]{adonis2}} function to allow ordinary formula input.
}
\examples{
data(caldana)
(pr <- permanova(compounds ~ light * time, caldana))
}
