% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hdaov}
\alias{plot.hdaov}
\title{SCI Plot for pairs that lead to rejection}
\usage{
\method{plot}{hdaov}(
  x,
  pair.no = NULL,
  pairs = NULL,
  sci.coordinate = NULL,
  hline.color = "black",
  sci.width = 0.5,
  sci.size = 1,
  sci.color = "blue",
  rej.point.color = "red",
  rej.point.size = 4,
  title = "Simultaneous Confidence Interval Plot",
  subtitle = NULL,
  xlabel = "Coordinates",
  ylabel = "Confidence Interval",
  ...
)
}
\arguments{
\item{x}{a list of matrices generated from \code{hdtest}.}

\item{pair.no}{a vector of integers where the integers denote the orders in whole pairs in analysis; only these confidence intervals of pairs specified in \code{pair.no} or \code{pairs} will be plotted; default value: \code{NULL}, which is selected from \code{pairs}.}

\item{pairs}{a matrix with two columns where each row specifies a pair of populations in analysis; only these confidence intervals of pairs specified in \code{pair.no} or \code{pairs} will be plotted; default value: \code{NULL}, so that the specified pairs for plotting SCI are all pairs that lead to rejection.}

\item{sci.coordinate}{a vector of integers where each integer specifies the coordinate of SCI; only these specified coordinates of SCI will be plotted; default value: \code{NULL}, which is equivalent to all coordinates.}

\item{hline.color}{the color of horizontal line in plot; the color setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{'black'}.}

\item{sci.width}{the width of whisker in confidence interval of each coordinate; the width setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{0.5}.}

\item{sci.size}{the thickness of line in confidence interval of each coordinate in plot; the size setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{1}.}

\item{sci.color}{the color of line in confidence interval of each coordinate in plot; the color setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{'blue'}.}

\item{rej.point.color}{the color of the points used to mark the coordinates of which confidence interval not containing 0; the color setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{'red'}.}

\item{rej.point.size}{the diameter of the points used to mark the coordinates of which confidence interval not containing 0; the color setting is the same as that in aesthetic specifications of \pkg{gglot2}; default value: \code{4}.}

\item{title}{the text for the title for plot; the title setting is the same as that in labels of \pkg{gglot2}; default value: \code{'Simultaneous Confidence Interval Plot'}.}

\item{subtitle}{the text for the subtitle for plot; the subtitle setting is the same as that in labels of \pkg{gglot2}; default value: \code{NULL} which indicates the pair number and group numbers.}

\item{xlabel}{the text for X coordinate for plot; the label setting is the same as that in labels of \pkg{gglot2}; default value: \code{'Coordinates'}.}

\item{ylabel}{the text for Y coordinate for plot; the label setting is the same as that in labels of \pkg{gglot2}; default value: \code{'Confidence Interval'}.}

\item{...}{other arguments passed to plot.}
}
\value{
plots of the SCIs of specified pairs
}
\description{
Plot the confidence interval of there pairs that lead to rejection after high-dimensional data analysis.
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) MASS::mvrnorm(30,rep(0.3*g,10),diag((1:10)^(-0.5*g))))

# test for the equality of mean vectors with pairs={(1,3),(2,4)}
plot(hdtest(X,alpha=0.05,pairs=matrix(1:4,2,2),tau=c(0.4,0.5,0.6), return.sci=TRUE))
}
\references{
\insertRef{Lopes2020}{hdanova}

\insertRef{Lin2020}{hdanova}
}
