% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdanova.R
\name{size.tau}
\alias{size.tau}
\title{Empirical Size Associated with Decay Parameter}
\usage{
size.tau(
  X,
  tau,
  alpha = 0.05,
  B = ceiling(50/alpha),
  pairs = NULL,
  verbose = F,
  method = "MGB",
  R = ceiling(10/alpha),
  ncore = 1,
  cuda = T,
  nblock = 32,
  tpb = 64,
  seed = sample.int(2^30, 1)
)
}
\arguments{
\item{X}{a matrix (one-sample) or a list of matrices (multiple-samples), with each row representing an observation.}

\item{tau}{real number(s) in the interval \code{[0,1)} for which the empirical size will be evaluated.}

\item{alpha}{significance level; default value: 0.05.}

\item{B}{the number of bootstrap replicates; default value: \code{ceiling(50/alpha)}.}

\item{pairs}{a matrix with two columns, only used when there are more than two populations, where each row specifies a pair of populations for which the SCI is constructed; default value: \code{NULL}, so that SCIs for all pairs are constructed.}

\item{verbose}{TRUE/FALSE, indicator of whether to output diagnostic information or report progress; default value: FALSE.}

\item{method}{the evaluation method tau; possible values are 'MGB' (default), 'MGBA', 'RMGB', 'RMGBA', 'WB' and 'WBA' (see \code{\link{hdsci}} for details).}

\item{R}{the number of iterations; default value: \code{ceiling(25/alpha)}.}

\item{ncore}{the number of CPU cores to be used; default value: 1.}

\item{cuda}{T/F to indicate whether to use CUDA GPU implementation when the package \code{hdanova.cuda} is installed. This option takes effect only when \code{ncore=1}.}

\item{nblock}{the number of block in CUDA computation}

\item{tpb}{number of threads per block; the maximum number of total number of parallel GPU threads is then \code{nblock*tpb}}

\item{seed}{the seed for random number generator}
}
\value{
a vector of empirical size corresponding to \code{tau}.
}
\description{
Find the empirical size associated with the decay parameter conditional on a dataset
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) MASS::mvrnorm(30,rep(0.3*g,10),diag((1:10)^(-0.5*g))))

size.tau(X,tau=seq(0,1,by=0.1),alpha=0.05,pairs=matrix(1:4,2,2),R=100)
}
\references{
\insertRef{Lopes2020}{hdanova}

\insertRef{Lin2020}{hdanova}
}
