% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hdaov}
\alias{summary.hdaov}
\title{Produce summary of HD ANOVA}
\usage{
\method{summary}{hdaov}(object)
}
\arguments{
\item{object}{the object returned from \code{hdtest}, \code{hdsci} or \code{fdtest}}
}
\value{
An object with information on p-value and/or SCI
}
\description{
Produce summary of HD ANOVA
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) MASS::mvrnorm(30,rep(0.3*g,10),diag((1:10)^(-0.5*g))))

# test for the equality of mean vectors with pairs={(1,3),(2,4)}
summary(hdtest(X,alpha=0.05,pairs=matrix(1:4,2,2),tau=c(0.4,0.5,0.6)))
}
