% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update.without.factor}
\alias{update.without.factor}
\title{Update a Model without Factor}
\usage{
\method{update}{without.factor}(model, fac, hierarchical = TRUE)
}
\description{
Update a Model without Factor
}
