% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apca.R
\name{apca}
\alias{apca}
\title{ANOVA Principal Component Analysis - APCA}
\usage{
apca(formula, data, ...)
}
\arguments{
\item{formula}{Model formula accepting a single response (block) and predictors.}

\item{data}{The data set to analyse.}

\item{...}{Additional parameters for the asca_fit function.}
}
\value{
An object of class \code{apca}, inheriting from the general \code{asca} class.
TODO: Link to ASCA documentation, add reference
}
\description{
APCA function for fitting ANOVA Principal Component Analysis models.
}
\examples{
data(candies)
ap <- apca(assessment ~ candy, data=candies)
scoreplot(ap)
}
