% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDANOVA-package.R
\docType{package}
\name{HDANOVA}
\alias{HDANOVA-package}
\alias{HDANOVA}
\title{HDANOVA}
\description{
Included so far:
\itemize{
\item Basic ASCA+ with some extras and plotting
\item PC-ANOVA
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/khliland/HDANOVA/}
  \item Report bugs at \url{https://github.com/khliland/HDANOVA/issues/}
}

}
\author{
\strong{Maintainer}: Kristian Hovde Liland \email{kristian.liland@nmbu.no} (\href{https://orcid.org/0000-0001-6468-9423}{ORCID})

}
\keyword{internal}
