% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{hdsci}
\alias{hdsci}
\title{Simultaneous confidence interval (SCI) for high-dimensional data}
\usage{
hdsci(X, alpha = 0.05, side = "both", tau = NULL, B = 1000,
  pairs = NULL, Sig = NULL, verbose = F)
}
\arguments{
\item{X}{a matrix (one sample) or a list of matrices (two samples or more), with observations contained in rows}

\item{alpha}{significance level}

\item{side}{either of \code{c('lower','upper', or 'both')}. default value is 'both'}

\item{tau}{the decay parameter, automatically selected if set to \code{NULL}}

\item{B}{the number of bootstrap replicates}

\item{pairs}{a vector with two columns. When number of samples is larger than two, each row specifies the pairs of samples for which the SCI are constructed}

\item{Sig}{a matrix (one sample) or a list of matrices, each of them is the covariance matrix of a sample}
}
\value{
a list of the following objects: \code{sci.lower} (\code{sci.upper}) is a vector (one- or two-sample) or a list of vectors (three or more samples) specifying the lower (upper) bound of SCI for the mean (one sample) or the difference of means of each pair of samples; when number of samples is larger than two, \code{pairs} is a matrix of two columns, each row containing the a pair of indices of samples whose SCI is constructed
}
\description{
Construct (1-\code{alpha}) simultaneous confidence interval (SCI)  for the mean or difference of means of high-dimensional vectors
}
\references{
\insertRef{Lopes2019+}{hdanova}

\insertRef{Lin2020}{hdanova}
}
