% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{hdtest}
\alias{hdtest}
\title{Hypothesis test for high-dimensional data}
\usage{
hdtest(X, alpha = 0.05, tau = NULL, B = 1000, pairs = NULL,
  Sig = NULL, verbose = F)
}
\arguments{
\item{X}{a matrix (one sample) or a list of matrices (two samples or more), with observations contained in rows}

\item{alpha}{significance level}

\item{tau}{the decay parameter, automatically selected if set to \code{NULL}}

\item{B}{the number of bootstrap replicates}

\item{pairs}{a matrix with two columns, used when there are more than two populations, each row specifying a pair of populations whose means are compared, and if set to \code{NULL}, comparisons for all pairs are performed}

\item{Sig}{a matrix (one sample) or a list of matrices, each of them is the covariance matrix of a sample and automatically estimated if \code{NULL}}
}
\value{
a list of objects: \code{reject} is a boolean variable indicating whether the null is rejected or not. If the null if reject, \code{rej.pairs} optionally gives the pairs of samples that lead to rejection
}
\description{
Test the mean or differences of means of high-dimensional vectors are zero or not
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) MASS::mvrnorm(30,rep(0,10),diag((1:10)^(-0.5*g))))

# test for the equality of mean vectors with pairs={(1,3),(2,4)}
hdtest(X,alpha=0.05,pairs=matrix(1:4,2,2))$reject
}
\references{
\insertRef{Lopes2019+}{hdanova}

\insertRef{Lin2020}{hdanova}
}
