% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{hdtest}
\alias{hdtest}
\title{Hypothesis test for high-dimensional data}
\usage{
hdtest(X, alpha = 0.05, tau = NULL, B = 1000, pairs = NULL,
  Sig = NULL, verbose = F)
}
\arguments{
\item{X}{a matrix (one sample) or a list of matrices (two samples or more), with observations contained in rows}

\item{alpha}{significance level}

\item{tau}{the decay parameter, automatically selected if set to \code{NULL}}

\item{B}{the number of bootstrap replicates}

\item{pairs}{a vector with two columns. When number of samples is larger than two, each row specifies the pairs of samples for which the SCI are constructed}

\item{Sig}{a matrix (one sample) or a list of matrices, each of them is the covariance matrix of a sample}
}
\value{
a list of objects: \code{reject} is a boolean variable indicating whether the null is rejected or not. If the null if reject, \code{rej.pairs} optionally gives the pairs of samples that lead to rejection
}
\description{
Test the mean or differences of means of high-dimensional vectors are zero or not
}
\references{
\insertRef{Lopes2019+}{hdanova}

\insertRef{Lin2020}{hdanova}
}
