% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcanova.R
\name{pcanova}
\alias{pcanova}
\alias{print.pcanova}
\alias{summary.pcanova}
\title{Principal Components Analysis of Variance Simultaneous Component Analysis - PC-ANOVA}
\usage{
pcanova(formula, data, ncomp = 0.9, ...)

\method{print}{pcanova}(x, ...)

\method{summary}{pcanova}(object, ...)
}
\arguments{
\item{formula}{Model formula accepting a single response (block) and predictor names separated by + signs.}

\item{data}{The data set to analyse.}

\item{ncomp}{The number of components to retain, proportion of variation or default = minimum cross-validation error.}

\item{weights}{Optional object weights.}

\item{subset}{Subset of objects}

\item{na.action}{How to handle NAs (no action implemented).}

\item{family}{Error distributions and link function for Generalized Linear Models.}

\item{SStype}{Sums-of-squares type for Analysis of Variance (I/II/III), defaults to "III".}
}
\value{
TODO: Update. A \code{pcanova} object containing loadings, scores, explained variances, etc. The object has
associated plotting (\code{\link{pcanova_plots}}) and result (\code{\link{pcanova_results}}) functions.
}
\description{
This is a quite general and flexible implementation of PC-ANOVA.
}
\details{
TODO: Update details. PC-ANOVA is a method which decomposes a multivariate response according to one or more design
variables. ANOVA is used to split variation into contributions from factors, and PCA is performed
on the corresponding least squares estimates, i.e., \code{Y = X1 B1 + X2 B2 + ... + E = T1 P1' + T2 P2' + ... + E}.
This version of PC-ANOVA encompasses variants of LiMM-PCA, generalized PC-ANOVA and covariates PC-ANOVA. It includes
confidence ellipsoids for the balanced fixed effect PC-ANOVA.
}
\examples{
# Load candies data
data(candies)

# Basic PC-ANOVA model with two factors, cross-validated opt. of #components
mod <- pcanova(assessment ~ candy + assessor, data = candies)
print(mod)

# PC-ANOVA model with interaction, minimum 90\% explained variance
mod <- pcanova(assessment ~ candy * assessor, data = candies, ncomp = 0.9)
print(mod)
summary(mod)

# Tukey group letters for 'candy' per component
lapply(mod$models, function(x)
       mixlm::cld(mixlm::simple.glht(x,
                                     effect = "candy")))

# Result plotting
loadingplot(mod, scatter=TRUE, labels="names")
scoreplot(mod)

# Mixed Model PC-ANOVA, random assessor
mod.mix <- pcanova(assessment ~ candy + r(assessor), data=candies, ncomp = 0.9)
scoreplot(mod.mix)
# Fixed effects
summary(mod.mix)


# TODO, example with lme4?
# Random effects
lapply(mod.mix$models, lme4::ranef)

}
\references{
Luciano G, Næs T. Interpreting sensory data by combining principal
component analysis and analysis of variance. Food Qual Prefer. 2009;20(3):167-175.
}
\seealso{
TODO
}
