% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdanova.R
\name{size.tau}
\alias{size.tau}
\title{Empirical Size Associated with Decay Parameter}
\usage{
size.tau(
  X,
  tau,
  alpha = 0.05,
  B = ceiling(50/alpha),
  pairs = NULL,
  verbose = F,
  R = ceiling(25/alpha),
  method = "MGB"
)
}
\arguments{
\item{X}{a matrix (one-sample) or a list of matrices (multiple-samples), with each row representing an observation.}

\item{tau}{real number(s) in the interval \code{[0,1)} for which the empirical size will be evaluated.}

\item{alpha}{significance level; default value: 0.05.}

\item{B}{the number of bootstrap replicates; default value: \code{ceiling(50/alpha)}.}

\item{pairs}{a matrix with two columns, only used when there are more than two populations, where each row specifies a pair of populations for which the SCI is constructed; default value: \code{NULL}, so that SCIs for all pairs are constructed.}

\item{verbose}{TRUE/FALSE, indicator of whether to output diagnostic information or report progress; default value: FALSE.}

\item{R}{the number of iterations; default value: \code{ceiling(25/alpha)}.}

\item{method}{the evaluation method tau; possible values are 'MGB' (default), 'MGBA', 'RGB', 'RGBA', 'WB' and 'WBA' (see \code{\link{hdsci}} for details).}
}
\value{
a vector of empirical size corresponding to \code{tau}.
}
\description{
Find the empirical size associated with the decay parameter conditional on a dataset
}
\examples{
# simulate a dataset of 4 samples
X <- lapply(1:4, function(g) MASS::mvrnorm(30,rep(0.3*g,10),diag((1:10)^(-0.5*g))))

# test for the equality of mean vectors with pairs={(1,3),(2,4)}
size.tau(X,tau=seq(0,1,by=0.1),alpha=0.05,pairs=matrix(1:4,2,2),R=100)
}
\references{
\insertRef{Lopes2020}{hdanova}

\insertRef{Lin2020}{hdanova}
}
